Updating to work in
Updating to Python 3.7 to work in EDMC v3.5.1.0. Confirmed that it worked in EDMC v3.4.6.0 in Python 2.7.
Fixed imports, prints and chrs. chrs required use of a new expandedChr function, which uses unichr in Python 2.7 and uses chr in Python 3.7 for the expanding chat window symbol.
Haven't exhausted checked every bit of the code for edge cases, but was able to use it to grab chat in both versions.